refactor: simplify component
