+++ packages/shared/routes/onboarding/OnboardingCustomNetworkView.svelte
@@ -1,18 +1,61 @@
 <script lang="typescript">
     import { OnboardingLayout, Text, Button, Spinner, NodeConfigurationForm } from 'shared/components'
     import { localize } from '@core/i18n'
-    import { INode } from '@core/network'
+    import { INode, NetworkType, validateAndCleanNodeData } from '@core/network'
     import { appRouter } from '@core/router'
+    import { destroyProfileManager, getNodeInfo } from '@core/profile-manager'
+    import { activeProfile, addNode, createNewProfile, newProfile } from '@core/profile'
+    import { showAppNotification } from '@lib/notifications'
 
     let nodeConfigurationForm: NodeConfigurationForm
     let node: INode
     let isBusy = false
+    let formError = ''
+
+    const profile = $newProfile ? newProfile : activeProfile
+    const clientOptions = $profile.settings?.clientOptions
 
     function onBackClick(): void {
         $appRouter.previous()
     }
-    function onSuccess(): void {
-        $appRouter.next()
+
+    async function onSuccess(): Promise<void> {
+        try {
+            await getNodeInfo(node.url)
+            $appRouter.next()
+        } catch (err) {
+            await destroyProfileManager()
+            formError = localize('error.node.invalid')
+        }
+    }
+
+    async function handleAddNode(): Promise<void> {
+        formError = ''
+        nodeConfigurationForm.validate()
+        if (!formError) {
+            isBusy = true
+            try {
+                if (!clientOptions) {
+                    const cleanedNode = validateAndCleanNodeData(node)
+                    await createNewProfile(
+                        $profile.isDeveloperProfile,
+                        $newProfile.networkProtocol,
+                        NetworkType.PrivateNet,
+                        cleanedNode
+                    )
+                } else {
+                    await addNode(node, profile)
+                }
+                onSuccess()
+            } catch (err) {
+                showAppNotification({
+                    type: 'error',
+                    message: localize(err?.error ?? 'error.global.generic'),
+                })
+            } finally {
+                isBusy = false
+            }
+        }
     }
 </script>
 
@@ -25,10 +68,10 @@
         <NodeConfigurationForm
             bind:this={nodeConfigurationForm}
             bind:node
-            bind:isBusy
+            bind:formError
+            {isBusy}
             hideButtons
             hideCheckbox
-            {onSuccess}
         />
     </div>
     <div slot="leftpane__action">
@@ -37,7 +80,7 @@
             type="submit"
             form="node-config-form"
             classes="w-full"
-            onClick={nodeConfigurationForm?.handleAddNode}
+            onClick={handleAddNode}
         >
             {#if isBusy}
                 <Spinner busy={isBusy} message={localize('popups.node.addingNode')} classes="justify-center" />
